% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_small.R
\name{add}
\alias{add}
\title{Package provides sample data that are loaded with package loading.}
\usage{
add(x, y)
}
\arguments{
\item{x}{and y are variables
#not RUN}
}
\description{
Package provides sample data that are loaded with package loading.
}
